Removed Install generator
